Updated to 1.8.9.9 Updated GRF Editor framework to .Net 4.8 and libraries to .Net Standard 2.0. Refactored the project configurations to be more friendly. Removed XDMessaging and zlib.net libraries. Bunch of other minor fixes.

+++ GRF/FileFormats/ThorFormat/ThorPacker.cs
@@ -183,7 +183,7 @@ namespace GRF.FileFormats.ThorFormat {
 
 					tableStream.Write(BitConverter.GetBytes(entry.TemporarySizeCompressedAlignment == 0 ? entry.SizeCompressed : entry.TemporarySizeCompressedAlignment), 0, 4);
 					tableStream.Write(BitConverter.GetBytes(entry.SizeDecompressed), 0, 4);
-					tableStream.Write(BitConverter.GetBytes(entry.TemporaryOffset + 12), 0, 4);
+					tableStream.Write(BitConverter.GetBytes((uint)entry.TemporaryOffset + 12), 0, 4);
 				}
 
 				tableStream.Seek(0, SeekOrigin.Begin);
